refactor: separate video player and title widgets
